DOT-60 Fixed Team enum

+++ Scenes/Weapons/Bows/SimpleBow.cs
@@ -27,7 +27,7 @@ public partial class SimpleBow : Bow
             deliverSound.Play();
             Node arrow = Arrow.Instantiate();
             Vector2 direction = (BowDirection.GlobalPosition - endOfBow.GlobalPosition).Normalized();
-            globals.EmitSignal(nameof(Globals.ArrowFired), arrow, team, endOfBow.GlobalPosition, direction);
+            globals.EmitSignal(nameof(Globals.ArrowFired), arrow, (int)team, endOfBow.GlobalPosition, direction);
             SetCurrentAmmo(CurrentAmmo - 1);
             attackCooldown.Start();
             Idle();
@@ -37,7 +37,7 @@ public partial class SimpleBow : Bow
             deliverSound.Play();
             Node arrow = FireArrow.Instantiate();
             Vector2 direction = (BowDirection.GlobalPosition - endOfBow.GlobalPosition).Normalized();
-            globals.EmitSignal(nameof(Globals.ArrowFired), arrow, team, endOfBow.GlobalPosition, direction);
+            globals.EmitSignal(nameof(Globals.ArrowFired), arrow, (int)team, endOfBow.GlobalPosition, direction);
             SetCurrentAmmo(CurrentAmmo - 1);
             attackCooldown.Start();
             Idle();
